Move total calc function into its own file
